Fix min distance displacement and catch db exceptions

+++ diy-geofence-java/src/main/java/com/ashwin/android/diygeofencejava/DatabaseHandler.java
@@ -6,7 +6,6 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-
 import java.util.HashSet;
 import java.util.Set;
 
@@ -62,42 +61,73 @@ public class DatabaseHandler extends SQLiteOpenHelper {
         values.put(COLUMN_LNG, lng);
         values.put(COLUMN_RAD, rad);
 
-        SQLiteDatabase db = this.getWritableDatabase();
-        long result = db.insertWithOnConflict(TABLE_GEOFENCE, null, values, SQLiteDatabase.CONFLICT_REPLACE);
-        db.close();
+        long result = -1;
+        try {
+            SQLiteDatabase db = this.getWritableDatabase();
+            result = db.insertWithOnConflict(TABLE_GEOFENCE, null, values, SQLiteDatabase.CONFLICT_REPLACE);
+            db.close();
+        } catch (Throwable t) {
+            Logger.e("Exception while adding geofence: " + id, t);
+        }
         return result != -1L;
     }
 
     boolean removeGeofence(String id) {
         boolean result = false;
         String query = "SELECT * FROM " + TABLE_GEOFENCE + " WHERE " + COLUMN_ID  + " = '" + id + "'";
-        SQLiteDatabase db = this.getWritableDatabase();
-        Cursor cursor = db.rawQuery(query, null);
-        if (cursor.moveToFirst()) {
-            db.delete(TABLE_GEOFENCE, COLUMN_ID + " = ?", new String[]{id});
-            cursor.close();
-            result = true;
+        try {
+            SQLiteDatabase db = this.getWritableDatabase();
+            Cursor cursor = db.rawQuery(query, null);
+            if (cursor != null && cursor.moveToFirst()) {
+                db.delete(TABLE_GEOFENCE, COLUMN_ID + " = ?", new String[]{id});
+                cursor.close();
+                result = true;
+            }
+
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
+
+            db.close();
+
+        } catch (Throwable t) {
+            Logger.e("Exception while removing geofence: " + id, t);
         }
-        db.close();
         return result;
     }
 
     boolean removeAllGeofences() {
-        SQLiteDatabase db = this.getWritableDatabase();
-        db.delete(TABLE_GEOFENCE, null, null);
-        return true;
+        try {
+            SQLiteDatabase db = this.getWritableDatabase();
+            db.delete(TABLE_GEOFENCE, null, null);
+            db.close();
+            return true;
+        } catch (Throwable t) {
+            Logger.e("Exception while removing all geofences", t);
+        }
+        return false;
     }
 
     DiyGeofenceData getGeofence(String id) {
         DiyGeofenceData result = null;
         String query = "SELECT * FROM " + TABLE_GEOFENCE + " WHERE " + COLUMN_ID + " = '" + id + "'";
-        SQLiteDatabase db = this.getWritableDatabase();
-        Cursor cursor = db.rawQuery(query, null);
-        if (cursor.moveToFirst()) {
-            double lat = cursor.getDouble(cursor.getColumnIndex(COLUMN_LAT));
-            double lng = cursor.getDouble(cursor.getColumnIndex(COLUMN_LNG));
-            double rad = cursor.getDouble(cursor.getColumnIndex(COLUMN_RAD));
-            result = new DiyGeofenceData(id, lat, lng, rad);
+        try {
+            SQLiteDatabase db = this.getWritableDatabase();
+            Cursor cursor = db.rawQuery(query, null);
+            if (cursor != null && cursor.moveToFirst()) {
+                double lat = cursor.getDouble(cursor.getColumnIndex(COLUMN_LAT));
+                double lng = cursor.getDouble(cursor.getColumnIndex(COLUMN_LNG));
+                double rad = cursor.getDouble(cursor.getColumnIndex(COLUMN_RAD));
+                result = new DiyGeofenceData(id, lat, lng, rad);
+            }
+
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
+
+            db.close();
+        } catch (Throwable t) {
+            Logger.e("Exception while getting geofence: " + id, t);
         }
         return result;
     }
@@ -105,19 +135,27 @@ public class DatabaseHandler extends SQLiteOpenHelper {
     Set<DiyGeofenceData> getAllGeofences() {
         Set<DiyGeofenceData> result = new HashSet<DiyGeofenceData>();
         String query = "SELECT * FROM " + TABLE_GEOFENCE;
-        SQLiteDatabase db = this.getWritableDatabase();
-        Cursor cursor = db.rawQuery(query, null);
-        if (cursor.moveToFirst()) {
-            do {
-                String id = cursor.getString(cursor.getColumnIndex(COLUMN_ID));
-                double lat = cursor.getDouble(cursor.getColumnIndex(COLUMN_LAT));
-                double lng = cursor.getDouble(cursor.getColumnIndex(COLUMN_LNG));
-                double rad = cursor.getDouble(cursor.getColumnIndex(COLUMN_RAD));
-                result.add(new DiyGeofenceData(id, lat, lng, rad));
-            } while (cursor.moveToNext());
+        try {
+            SQLiteDatabase db = this.getWritableDatabase();
+            Cursor cursor = db.rawQuery(query, null);
+            if (cursor != null && cursor.moveToFirst()) {
+                do {
+                    String id = cursor.getString(cursor.getColumnIndex(COLUMN_ID));
+                    double lat = cursor.getDouble(cursor.getColumnIndex(COLUMN_LAT));
+                    double lng = cursor.getDouble(cursor.getColumnIndex(COLUMN_LNG));
+                    double rad = cursor.getDouble(cursor.getColumnIndex(COLUMN_RAD));
+                    result.add(new DiyGeofenceData(id, lat, lng, rad));
+                } while (cursor.moveToNext());
+            }
+
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
+
+            db.close();
+        } catch (Throwable t) {
+            Logger.e("Exception while getting all geofences", t);
         }
-        cursor.close();
-        db.close();
         return result;
     }
 }
